[C++] Add Win32 / POSIX compatibility layer for consistent handling of files over 2 GB
To properly do this, I had to make usages of `cstdio` "private" and push down the file interactions into a private implementation.

Per #155
Per #120

I've tested locally that this fixes the observed Windows failure on files > 2 GB. I will ask others to verify before I close the issues.

Author: Wes McKinney <wesm@apache.org>

Closes #187 from wesm/large-file-32-bit-issue and squashes the following commits:

309c0ce [Wes McKinney] Add missing include
0fddf40 [Wes McKinney] Refactor file handling to be cross-platform compatible on large files. Return error message when trying to read a large file on Windows 32-bit.

+++ cpp/src/feather/metadata.cc
@@ -421,7 +421,7 @@ size_t Table::num_columns() const {
   return table->columns()->size();
 }
 
-std::shared_ptr<Column> Table::GetColumn(size_t i) const {
+std::shared_ptr<Column> Table::GetColumn(int i) const {
   const fbs::CTable* table = static_cast<const fbs::CTable*>(table_);
   const fbs::Column* col = table->columns()->Get(i);
